+++ src/cli/dig.rs
@@ -3,9 +3,11 @@ mod commit_rewards;
 mod find_slot;
 mod launch;
 mod sync;
+mod sync_distributor;
 
 pub use clawback_rewards::*;
 pub use commit_rewards::*;
 pub use find_slot::*;
 pub use launch::*;
 pub use sync::*;
+pub use sync_distributor::*;
